requirements.txt
SpeechRecognition 
pyttsx3
pyaudio
ollama
openai
python-dotenv
requests
websockets
openwakeword
onnxruntime
faster-whisper
nvidia-cublas-cu12
nvidia-cudnn-cu12